requirements.txt
paho-mqtt==1.5
Adafruit-DHT==1.4.0
